requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
requests>=2.28.0
openai>=0.27.0
python-dotenv>=0.19.0
serpapi
urllib3
typing-extensions
streamlit-quill>=0.0.36
numpy
tqdm
beautifulsoup4>=4.11.0
sendgrid>=6.9.7
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5 